# simplify RANDAO recovery in `ShufflingRef` acceleration